Refactor theme-related variables and functions in proper file structures
