+++ src/App.js
@@ -17,51 +17,11 @@ import UserProvider from "./contexts/UserContext";
 import Profile from "./pages/Profile.js";
 import Reauthenticate from "./services/Reauthenticate.js";
 import FullPageTest from "./components/FullPageWrapper.js";
+import { getTheme, ThemeContext, themeReducer } from "./config/ThemeConfig.js";
 
-export const ThemeContext = createContext();
-const themeColors = [
-  "#e3bd8d",
-  "#ce5777",
-  "#41b3a3",
-  "#5f6caf",
-  "#E27d60",
-  "#557a95",
-  "#f7d1ba",
-  "#ff9a76",
-  "#f6f7d7",
-  "#d8e2dc",
-  "#e9c46a",
-  "#f4a261",
-  "#e76f51",
-  "#264653",
-  "#2a9d8f",
-  "#e9c46a",
-  "#f4a261",
-  "#e76f51",
-  "#264653",
-  "#2a9d8f",
-];
-function themeReducer(state, action) {
-  switch (action.type) {
-    case "SWITCH_THEME":
-      return (state + 1) % themeColors.length;
-    default:
-      throw new Error(`Unknown action: ${action.type}`);
-  }
-}
 const App = () => {
   const [themeColorIndex, dispatch] = useReducer(themeReducer, 0);
-  const theme = useMemo(
-    () =>
-      createTheme({
-        palette: {
-          primary: {
-            main: themeColors[themeColorIndex],
-          },
-        },
-      }),
-    [themeColorIndex],
-  );
+  const theme = useMemo(() => getTheme(themeColorIndex));
 
   return (
     <UserProvider>
@@ -86,10 +46,7 @@ const App = () => {
                   path="user/forgot-password"
                   element={<ForgotPassword />}
                 />
-                <Route
-                  path="user/reset-password"
-                  element={<ResetPassword />}
-                />
+                <Route path="user/reset-password" element={<ResetPassword />} />
               </Route>
             </Routes>
           </Router>
